Исправляет перерисовку компанента формы при изменении типа точки
