+++ src/const.js
@@ -1,49 +1,10 @@
-const pointTypes = [
-  {
-    title: 'Taxi',
-    value: 'taxi',
-  },
-  {
-    title: 'Bus',
-    value: 'bus',
-  },
-  {
-    title: 'Train',
-    value: 'train',
-  },
-  {
-    title: 'Ship',
-    value: 'ship',
-  },
-  {
-    title: 'Drive',
-    value: 'drive',
-  },
-  {
-    title: 'Flight',
-    value: 'flight',
-  },
-  {
-    title: 'Check-in',
-    value: 'check-in',
-  },
-  {
-    title: 'Sightseeing',
-    value: 'sightseeing',
-  },
-  {
-    title: 'Restaurant',
-    value: 'restaurant',
-  },
-];
-
 const SortType = {
   DAY: 'sort-day',
   TIME: 'sort-time',
   PRICE: 'sort-price',
 };
 
-const DateFormats = {
+const DateFormat = {
   DATE_FORMAT: 'YYYY-MM-DD',
   DATE_TIME_FORMAT: 'YYYY-MM-DDTHH:mm',
   DATE_TIME_FORM_POINTS: 'YY/MM/DD HH:mm',
@@ -52,4 +13,4 @@ const DateFormats = {
 };
 
 
-export {pointTypes, SortType, DateFormats};
+export {SortType, DateFormat};
